started reworking partitioners introducing factories
