+++ connector/src/main/scala/io/github/dejarol/azure/search/spark/connector/read/config/ReadConfig.scala
@@ -6,7 +6,7 @@ import io.github.dejarol.azure.search.spark.connector.core.JavaScalaConverters
 import io.github.dejarol.azure.search.spark.connector.core.config.{ExtendableConfig, SearchConfig, SearchIOConfig}
 import io.github.dejarol.azure.search.spark.connector.core.utils.SearchClients
 import io.github.dejarol.azure.search.spark.connector.read.filter.{ODataExpression, ODataExpressions}
-import io.github.dejarol.azure.search.spark.connector.read.partitioning.{DefaultPartitioner, SearchPartition, SearchPartitioner}
+import io.github.dejarol.azure.search.spark.connector.read.partitioning.{DefaultPartitioner, PartitionerFactory, RangePartitionerFactory, SearchPartition, SearchPartitioner}
 import org.apache.spark.sql.catalyst.util.CaseInsensitiveMap
 import org.apache.spark.sql.types.StructType
 
@@ -108,6 +108,18 @@ case class ReadConfig(override protected val options: CaseInsensitiveMap[String]
       classOf[DefaultPartitioner].asInstanceOf[Class[SearchPartitioner]],
       s => Class.forName(s).asInstanceOf[Class[SearchPartitioner]]
     )
+  }
+
+  def partitionerFactory: PartitionerFactory = {
+
+    get(ReadConfig.PARTITIONER_CLASS_CONFIG) match {
+      case Some(value) => value match {
+        case "range" => RangePartitionerFactory
+        case "faceted" => null
+        case _ => null
+      }
+      case None => DefaultPartitioner.FACTORY
+    }
   }
 
   /**
